Add bare bones express app
This commit contains the first bit of code we will need for our Node.js/Express.js web-app. It will display our project name on launch. This will be the only code pushed to the master branch as it is very simple and a building block to the rest of our web-app.