Add authentication and logout
I've changed the web-app to require authentication from the get-go. The oauth2.js file is from google's oauth2 tutorials. It handles google authentication and extracts information from the user's google account. We will use the profile.id as a unique identifier for all of our site's users and we will implement a function to store the user in our database on their very first log in. Lastly, a generic logout was added to the base page to allow users to sign out.